Fix crashes while creating room v12 rooms (#87)
The problem is that the server would reject the power level
specification and return an error, but seemingly still create the room.

We also would seemingly swallow this error? Sometimes? I'm not quite
sure.

Ultimately this resulted in a situation where all subsequent(?)
reconciliation runs would crash because they expected to find an
m.room.power_levels event already in the room, *always*, but for these
broken rooms, there would be none, and we would throw an error.

See the third bullet point of the v12 client notes at
https://spec.matrix.org/unstable/rooms/v12/.